Package rename and minor review fixes
