+++ spring-demo-classic/pom.xml
@@ -102,8 +102,8 @@
                 </configuration>
             </plugin>
 
-            <!-- The Jetty plugin allows us to easily test the development build by
-				running jetty:run on the command line. -->
+            <!-- The Jetty plugin allows us to easily test the development build by running jetty:run
+            on the command line. It also starts a web container to perform integration tests-->
             <plugin>
                 <groupId>org.eclipse.jetty</groupId>
                 <artifactId>jetty-maven-plugin</artifactId>
