+++ spring-demo-classic/src/main/java/com/vaadin/spring/framework7/samples/TrivialNavigationComponent.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.spring.tutorial;
+package com.vaadin.spring.framework7.samples;
 
 import com.vaadin.navigator.Navigator;
 import com.vaadin.navigator.View;
