+++ spring-demo-classic/src/main/webapp/WEB-INF/applicationContext.xml
@@ -6,6 +6,6 @@
                            http://www.springframework.org/schema/context
                            http://www.springframework.org/schema/context/spring-context-4.1.xsd">
 
-    <bean class="com.vaadin.spring.tutorial.NavigableSpringTestUI.SpringConfiguration" />
-    <context:component-scan base-package="com.vaadin.spring.tutorial" />
+    <bean class="com.vaadin.spring.framework7.samples.NavigableSpringTestUI.SpringConfiguration" />
+    <context:component-scan base-package="com.vaadin.spring.framework7.samples" />
 </beans>
